pyproject.toml
[project]
name = "comfyui-gemini-flash"
version = "0.1.0"
description = "A ComfyUI custom node that integrates Google's Gemini Flash 2.0 model"
authors = [
    { name = "Shmuel Ronen" },
    { name = "Jerrod Linderman -update to flash 2.0" }
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "google-genai",
    "pillow",
    "torchaudio"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit"
]

[tool.comfy-nodes]
publisher = "@nvmaxx"